+++ RELEASENOTES.md
@@ -2,7 +2,9 @@ These are the release notes for the [mediawiki-api](README.md).
 
 ## Version 0.2 (under development)
 
-* ...
+* Altered everything for changed in mediawiki-datamodel
+* Removed Edit << action class
+* Introcudes NewEditInfo and NewRevision
 
 ## Version 0.1 (2014-02-23)